routes/account.js: support case insensitive account requests
Always use lowercase account in requests.

When accessing /account/:account directly using mixed case
checksummed addresses, as is done when following
"Transaction History" links from MyEtherWallet, users encounter
a page with a correct account balance but showing no transactions.

The correct balance shows up because the account is converted
to lowercase before that query.

Instead, convert the account from the request to lowercase and
use that value for all queries on the page.

Signed-off-by: Michael Ira Krufky <mkrufky@gmail.com>